move PD scripts into scripts/population_diversity/
